refactor(router): split 1712-line tests.rs into 7 focused sub-files
The router/tests.rs test suite (1712 lines) is split into a directory
of focused test modules, each under 350 lines:

- tests.rs: shared re-exports and two state-builder helpers (~249 lines)
- tests/handshake.rs: 6 handshake/initialization tests
- tests/gc.rs: MockAgent/NonLgtmAgent stubs + 3 GC adopt tests
- tests/metrics_rules.rs: 3 rule-check and metrics tests
- tests/events.rs: 4 event log/query tests
- tests/skills.rs: 3 skill CRUD tests
- tests/threads_and_misc.rs: 7 thread, turn, stats, and learn tests
- tests/exec_plan.rs: 5 ExecPlan persistence tests

No test behavior changes; all 695 server tests continue to pass.
